Rust API gateway, part 3 of N (#418)
- Implement create_user

I use a slightly different EP3 user provisioning scheme as compared to
the Python version.

First, ids in EP3 will now be ULIDs, which are also byte-compatible as
UUIDs. The following correspondences will hold:

- EP3 user ID == user.id
- EP3 display name == user.username

On startup, we load all EP3 users and compare it to what we know in our
database. For now, it just emits a warning; in prod, we may choose to
error.

On create_user, we look for the corresponding EP3 user to see if
user.username or EP3 display name already exists--if so, we just link to
that EP3 user. If not, we will provision a new EP3 user along the same
lines as the Python code.